feat: refine folder structure so that api doc generated by typedoc has good structure
