+++ bindings/javascript/src/types/enums.ts
