+++ bindings/javascript/typedoc.json
@@ -1,6 +1,6 @@
 {
     "$schema": "https://typedoc.org/schema.json",
-    "entryPoints": ["src/database", "src/enums.ts", "src/libredwg.ts"],
+    "entryPoints": ["src/database", "src/types", "src/libredwg.ts"],
     "entryPointStrategy": "expand", 
     "name": "MLightCAD LibreDwg-Web API Docs",
     "out": "docs",
